Add files via upload
A/B tests are very commonly performed by data analysts and data scientists. For this project, you will be working to understand the results of an A/B test run by an e-commerce website.  
the goal is to work through this notebook to help the company understand if they should:
- Implement the new webpage, 
- Keep the old webpage, or 
- Perhaps run the experiment longer to make their decision.
This project demonstrates the understanding and application of three methods to perform the same task where these methods confirm the results obtained and offer extra exploration options
The methods used are
1- Hypothesis testing
2- Z-Testing
3- Logistic Regression